refactor: view단에 있는 controller 코드 롤백
